Refactor daemon transport negotiation into modules
