+++ crates/transport/src/daemon/tests/mod.rs
@@ -1,0 +1,3 @@
+mod common;
+mod mapping;
+mod negotiation;
