+++ crates/transport/src/daemon/types/mod.rs
@@ -1,0 +1,5 @@
+mod handshake;
+mod parts;
+
+pub use handshake::LegacyDaemonHandshake;
+pub use parts::LegacyDaemonHandshakeParts;
